Replace interests overlay with inline multi-select
